Move VGSLoggingConfiguration to separate file
